setup vitest & add some basic UTs
