+++ packages/backend/package.json
@@ -8,7 +8,8 @@
         "dev:watch": "yarn generate:types && tsc-watch --preserveWatchOutput --onSuccess \"yarn dev --configPath ../../config.json --cacheDir ../../.sourcebot\"",
         "dev": "export PATH=\"$PWD/../../bin:$PATH\" && export CTAGS_COMMAND=ctags && node ./dist/index.js",
         "build": "yarn generate:types && tsc",
-        "generate:types": "tsx tools/generateTypes.ts"
+        "generate:types": "tsx tools/generateTypes.ts",
+        "test": "vitest --config ./vitest.config.ts"
     },
     "devDependencies": {
         "@types/argparse": "^2.0.16",
@@ -17,7 +18,8 @@
         "json-schema-to-typescript": "^15.0.2",
         "tsc-watch": "^6.2.0",
         "tsx": "^4.19.1",
-        "typescript": "^5.6.2"
+        "typescript": "^5.6.2",
+        "vitest": "^2.1.4"
     },
     "dependencies": {
         "@gitbeaker/rest": "^40.5.1",
